requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.0
requests>=2.26.0
python-multipart>=0.0.5
aiofiles>=0.7.0
tenacity>=8.0.0

# Model dependencies
llama-cpp-python>=0.2.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99

# Logging and monitoring
structlog>=21.5.0
prometheus-client>=0.12.0
python-json-logger>=2.0.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.16.0
pytest-cov>=2.12.1
httpx>=0.23.0

# Development
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910
